Combine common GTK toolbar code.
